initial draft for resolving connection error
