+++ package/kedro_viz/launchers/utils.py
@@ -1,0 +1,62 @@
+from time import sleep, time
+from typing import Any, Callable
+
+import requests
+
+
+class WaitForException(Exception):
+    """WaitForException: if func doesn't return expected result within the specified time"""
+
+
+def wait_for(
+    func: Callable,
+    expected_result: Any = True,
+    timeout: int = 60,
+    print_error: bool = True,
+    sleep_for: int = 1,
+    **kwargs,
+) -> None:
+    """
+    Run specified function until it returns expected result until timeout.
+
+    Args:
+        func (Callable): Specified function
+        expected_result (Any): result that is expected. Defaults to None.
+        timeout (int): Time out in seconds. Defaults to 10.
+        print_error (boolean): whether any exceptions raised should be printed.
+            Defaults to False.
+        sleep_for (int): Execute func every specified number of seconds.
+            Defaults to 1.
+        **kwargs: Arguments to be passed to func
+
+    Raises:
+         WaitForException: if func doesn't return expected result within the
+         specified time
+
+    """
+    end = time() + timeout
+
+    while time() <= end:
+        try:
+            retval = func(**kwargs)
+        except Exception as err:  # pylint: disable=broad-except
+            if print_error:
+                logger.error(err)
+        else:
+            if retval == expected_result:
+                return None
+        sleep(sleep_for)
+
+    raise WaitForException(
+        f"func: {func}, didn't return {expected_result} within specified timeout: {timeout}"
+    )
+
+
+def check_viz_up(host: str, port: int):  # pragma: no cover
+    url = f"http://{host}:{port}"
+    try:
+        response = requests.get(url, timeout=10)
+    except requests.ConnectionError:
+        return False
+
+    return response.status_code == 200
